Use size_t instead of uint16_t

+++ bm8563.h
@@ -97,8 +97,8 @@ extern "C" {
 
 /* These should be provided by the HAL. */
 typedef struct {
-    int (* read)(void *handle, uint8_t address, uint8_t reg, uint8_t *buffer, uint16_t size);
-    int (* write)(void *handle, uint8_t address, uint8_t reg, const uint8_t *buffer, uint16_t size);
+    int (* read)(void *handle, uint8_t address, uint8_t reg, uint8_t *buffer, size_t size);
+    int (* write)(void *handle, uint8_t address, uint8_t reg, const uint8_t *buffer, size_t size);
     void *handle;
 } bm8563_t;
